Make generic keyboard entry helper
